+++ js/chapters/chapter1.js
@@ -11,11 +11,10 @@ September 02, 2017
 */
 var story = new Array();
 
-const current_Chapter = 1;//TODO: remove this
 user.last_chapter_played = 1;
 pushVariablesToDB();
 
-const c = 19;
+const c = 19;//TODO: update this because it should be 20
 
 function Chapter1()
 {